+++ src/sdk/sdk.ts
@@ -11,13 +11,14 @@ import { Factory, PaymasterApi, SdkOptions } from './interfaces';
 import { Network } from "./network";
 import { BatchUserOpsRequest, Exception, getGasFee, MODULE_TYPE, onRampApiKey, openUrl, UserOperation, UserOpsRequest } from "./common";
 import { BigNumber, BigNumberish, Contract, TypedDataField, ethers, providers } from 'ethers';
-import { Networks, onRamperAllNetworks } from './network/constants';
+import { DEFAULT_QUERY_PAGE_SIZE, Networks, onRamperAllNetworks } from './network/constants';
 import { EtherspotWalletAPI, HttpRpcClient, VerifyingPaymasterAPI } from './base';
 import { TransactionDetailsForUserOp, TransactionGasInfoForUserOp } from './base/TransactionDetailsForUserOp';
 import { OnRamperDto, SignMessageDto, validateDto } from './dto';
 import { ErrorHandler } from './errorHandler/errorHandler.service';
 import { EtherspotBundler } from './bundler';
 import { ModularEtherspotWallet } from './contracts/src/ERC7579/wallet';
+import { ModuleInfo } from './base/EtherspotWalletAPI';
 
 /**
  * Modular-Sdk
@@ -77,7 +78,6 @@ export class ModularSdk {
 
     if (entryPointAddress == '') throw new Exception('entryPointAddress not set on the given chain_id')
     if (walletFactoryAddress == '') throw new Exception('walletFactoryAddress not set on the given chain_id')
-
     this.etherspotWallet = new EtherspotWalletAPI({
       provider,
       walletProvider: walletConnectProvider ?? walletProvider,
@@ -260,6 +260,10 @@ export class ModularSdk {
     return this.etherspotWallet._getAccountContract();
   }
 
+  async isModuleInstalled(moduleTypeId: MODULE_TYPE, module: string): Promise<boolean> {
+    return this.etherspotWallet.isModuleInstalled(moduleTypeId, module);
+  }
+
   async installModule(moduleTypeId: MODULE_TYPE, module: string, initData?: string): Promise<string> {
     const installData = await this.etherspotWallet.installModule(moduleTypeId, module, initData);
 
@@ -275,6 +279,18 @@ export class ModularSdk {
     return uoHash;
   }
 
+  async getPreviousModuleAddress(moduleTypeId: MODULE_TYPE, module: string): Promise<string> {
+    return this.etherspotWallet.getPreviousAddress(module, moduleTypeId);
+  }
+
+  async generateModuleDeInitData(moduleTypeId: MODULE_TYPE, module: string, moduleDeInitData: string): Promise<string> {
+    return await this.etherspotWallet.generateModuleDeInitData(moduleTypeId, module, moduleDeInitData);
+  }
+
+  async getPreviousAddress(moduleTypeId: MODULE_TYPE, targetAddress: string): Promise<string> {
+    return await this.etherspotWallet.getPreviousAddress(targetAddress, moduleTypeId);
+  }
+
   async uninstallModule(moduleTypeId: MODULE_TYPE, module: string, deinitData: string): Promise<string> {
     const uninstallData = await this.etherspotWallet.uninstallModule(moduleTypeId, module, deinitData);
 
@@ -288,6 +304,11 @@ export class ModularSdk {
     const op = await this.estimate();
     const uoHash = await this.send(op);
     return uoHash;
+  }
+
+  async getAllModules(pageSize: number = DEFAULT_QUERY_PAGE_SIZE): Promise<ModuleInfo> {
+    const modules = await this.etherspotWallet.getAllModules(pageSize);
+    return modules;
   }
 
   async totalGasEstimated(userOp: UserOperation): Promise<BigNumber> {